changed border of navbar button
there's this annoying issue where the navbar is made of an icon, and due to this, it's colour cannot be changed from pure black. ie, i cannot set it's colour to be a global variable, which is a bit annoying. So far, I made it so the border is adjusted but the icon still isn't *shruggie*